Create Imageprocess.java
I thought I uploaded this but it only went as far as my page on github. I don't know how to get the file here, so I created a new file and copy-pasted the text in. OK so I tried to change Brian's code as little as possible, so when I did change something I mostly commented it out so he can see what I changed. However, the pixels are not lining up correctly because I got results but not the ones I expected, and the date was reading as 31 and the sheet as 9.